% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{densityplot2d}
\alias{densityplot2d}
\title{Plots a 2D Density}
\usage{
densityplot2d(x, y, nbins = 50, max.points = 5,
  colfunc = colorRampPalette(c("lightblue", "green", "yellow", "red")), ...)
}
\arguments{
\item{x}{vector containing the X coordinates of the points.}

\item{y}{vector containing the Y coordinates of the points.
which are passed on to the plotting function so the plot can be
customized.}
}
\description{
This function tries to replicate 'smart' 2D density plots 
as they are found in flow cytometry data analysis software.
These plots commonly work with millions of points. In dense
regions, they will plot these as a heatmap; in sparse regions,
they will show the actual points.
}
