% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{densityplot2d}
\alias{densityplot2d}
\title{Plots a 2D Density}
\usage{
densityplot2d(x, y, nbins = 50, max.points = 5,
  colfunc = colorRampPalette(c("gray", "darkblue", "lightblue", "green",
  "yellow", "orange", "red")), tf.fun = identity, pch = 19, cex = 0.3,
  color.levels = 20, ...)
}
\arguments{
\item{x}{vector containing the X coordinates of the points.}

\item{y}{vector containing the Y coordinates of the points.
which are passed on to the plotting function so the plot can be
customized.}

\item{nbins}{how many bins to use for histogram.}

\item{max.points}{when to start plotting heatmap instead of 
individual points.}

\item{colfunc}{color mapping for heatmap.}

\item{tf.fun}{function to apply to histogram values before 
plotting them, \code{log} can be useful.}

\item{pch}{point character to use for individual points.}

\item{cex}{point character expansion to use for individual points.}

\item{color.levels}{how many different colors to use for heatmap.}

\item{...}{further arguments to be passed to the plotting function,
such as \code{xlim} or \code{xlab}.}
}
\description{
This function tries to replicate 'smart' 2D density plots 
as they are found in flow cytometry data analysis software.
These plots commonly work with millions of points. In dense
regions, they will plot these as a heatmap; in sparse regions,
they will show the actual points.
}
